mummi operator cpu experiments
I do not know why the mlserver generated far fewer samples
for cpu vs. gpu - the underlying orchestration should be
the same, but the resources are different. I need to look
at the code to understand why. It must be related to having
one more GPU being occupied, and due to the relative
percentages, the first step needing to ask for more
samples constantly.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>