mummi operator gpu: re-run with bug fixes
I spent time today figuring out why the generation was so strange
and figured it out. It was because of the communication between
traditional mummi and kubernetes - traditional mummi was asking
for queued jobs and anything inactive not successful (including
failed) was treated as that. When a job errored, or when another
job type was run, since we considered both failed and all jobs
running in the set that calculates the threshold, we would always
generate new samples (even if we did not need them)

Signed-off-by: vsoch <vsoch@users.noreply.github.com>