flux: add pulling times
For all experiments, pulling times are included, with the
exception of flux. To make the comparison fair, we add
the pulling time, which I recorded for 3x iterations, and
we would assume all nodes to pull at once (adding only
the time to pull each of the mlrunner, createsim, and cganalysis
containers to the system. Between sets of three I clean the
cache for a new pull, and subsequent pulls after the
mlrunner are allowed to use the singularity cache, as would
happen on an actual node

Signed-off-by: vsoch <vsoch@users.noreply.github.com>