add final cpu mummi runs
I also am renaming the previously cganalysis-fail to cganalysis.
It did not make sense to me that all jobs failed - these
were flagged as failed right before the push and it was erroneous.
In addition to an exit code of 0, an output file was looked for
in the wrong path. So these were erroneously labeled with a fail
tag when the jobs were ok. I have fixed the bug in the operator.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>