redo of iteration for gpu autoscaling
There are two runs of the previous gpu-autoscale-0 that
cut out 5-7 minutes early for cganalysis and it is not
clear why. I do not think we can use the data (I suspect
it is an edge case error for cganalysis that resulted
in still a zero exit code) so I did a rerun.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>